engine: extract esm/folder_browser.js
